refactor: Rename middleware to extensions
