+++ atsume/settings/default_settings.py
@@ -13,7 +13,7 @@ COMPONENT_DM_PERMISSIONS = []
 
 COMPONENT_GUILD_PERMISSIONS = {}
 
-MIDDLEWARE = []
+EXTENSIONS = []
 
 HIKARI_LOGGING = False
